Move docker operations to the same file
