+++ dev-requirements.txt
@@ -1,6 +1,7 @@
 git+https://github.com/jupyterhub/the-littlest-jupyterhub
 notebook
 pytest
+pytest-aiohttp
 pytest-asyncio
 pytest-cov
 requests-mock
