+++ tests/conftest.py
@@ -5,7 +5,6 @@ import sys
 import pytest
 
 from aiodocker import Docker, DockerError
-
 from jupyterhub.tests.conftest import (
     io_loop,
     event_loop,
@@ -13,6 +12,9 @@ from jupyterhub.tests.conftest import (
     pytest_collection_modifyitems,
 )
 from jupyterhub.tests.mocking import MockHub
+from tljh_repo2docker import Repo2DockerSpawner
+from tljh_repo2docker.builder import BuildHandler
+from tljh_repo2docker.images import ImagesHandler
 
 
 @pytest.fixture(scope='module')
@@ -42,10 +44,6 @@ def app(request, io_loop):
     Adapted from:
     https://github.com/jupyterhub/jupyterhub/blob/8a3790b01ff944c453ffcc0486149e2a58ffabea/jupyterhub/tests/conftest.py#L74
     """
-    from tljh_repo2docker import Repo2DockerSpawner
-    from tljh_repo2docker.builder import BuildHandler
-    from tljh_repo2docker.images import ImagesHandler
-
     mocked_app = MockHub.instance()
     mocked_app.spawner_class = Repo2DockerSpawner
     mocked_app.template_paths.insert(
